requirements.txt
asgiref==3.3.1
backcall==0.2.0
colorama==0.4.4
decorator==4.4.2
Django==3.1.4
django-crispy-forms==1.10.0
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
parso==0.7.1
phonenumberslite==8.12.14
pickleshare==0.7.5
prompt-toolkit==3.0.8
psycopg2==2.8.6
Pygments==2.7.3
pytz==2020.4
sqlparse==0.4.1
traitlets==5.0.5
wcwidth==0.2.5
